+++ lib/string/auth_string_extensions.dart
@@ -2,7 +2,7 @@ part of '../flutter_extend.dart';
 
 extension AuthStringExtensions on String {
   /// Check if Email is Valid
-  bool get xIsEmailValid {
+  bool get isEmailValid {
     final bool isEmailValid = RegExp(
             r"^[a-zA-Z0-9.a-zA-Z0-9.!#$%&'*+-/=?^_`{|}~]+@[a-zA-Z0-9]+\.[a-zA-Z]+")
         .hasMatch(this);