+++ lib/string/global_string_extensions.dart
@@ -2,11 +2,11 @@ part of '../flutter_extend.dart';
 
 extension GlobalStringExtensions on String {
   /// Last N Characters
-  String xGetLastNCharacters(int n) =>
+  String getLastNCharacters(int n) =>
       length >= n ? substring(length - n) : this;
 
   /// Truncate Name
-  String xTruncateName() {
+  String truncateName() {
     final wordArray = split(' ');
 
     return wordArray.length > 1